applied feedback
