+++ src/app/+admin/admin-search-page/admin-search-results/admin-search-result-list-element/collection-search-result/collection-admin-search-result-list-element.component.spec.ts
@@ -1,26 +1,26 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { NO_ERRORS_SCHEMA } from '@angular/core';
+import { CollectionAdminSearchResultListElementComponent } from './collection-admin-search-result-list-element.component';
 import { TranslateModule } from '@ngx-translate/core';
-import { TruncatableService } from '../../../truncatable/truncatable.service';
-import { CollectionElementLinkType } from '../../../object-collection/collection-element-link.type';
-import { ViewMode } from '../../../../core/shared/view-mode.model';
-import { Collection } from '../../../../core/shared/collection.model';
+import { TruncatableService } from '../../../../../shared/truncatable/truncatable.service';
+import { CollectionElementLinkType } from '../../../../../shared/object-collection/collection-element-link.type';
+import { ViewMode } from '../../../../../core/shared/view-mode.model';
+import { CollectionSearchResult } from '../../../../../shared/object-collection/shared/collection-search-result.model';
+import { Collection } from '../../../../../core/shared/collection.model';
 import { By } from '@angular/platform-browser';
 import { RouterTestingModule } from '@angular/router/testing';
-import { CommunityAdminSearchResultListElementComponent } from './community-admin-search-result-list-element.component';
-import { CommunitySearchResult } from '../../../object-collection/shared/community-search-result.model';
-import { getCommunityEditPath } from '../../../../+community-page/community-page-routing.module';
+import { getCollectionEditPath } from '../../../../../+collection-page/collection-page-routing.module';
 
-describe('CommunityAdminSearchResultListElementComponent', () => {
-  let component: CommunityAdminSearchResultListElementComponent;
-  let fixture: ComponentFixture<CommunityAdminSearchResultListElementComponent>;
+describe('CollectionAdminSearchResultListElementComponent', () => {
+  let component: CollectionAdminSearchResultListElementComponent;
+  let fixture: ComponentFixture<CollectionAdminSearchResultListElementComponent>;
   let id;
   let searchResult;
 
   function init() {
     id = '780b2588-bda5-4112-a1cd-0b15000a5339';
-    searchResult = new CommunitySearchResult();
+    searchResult = new CollectionSearchResult();
     searchResult.indexableObject = new Collection();
     searchResult.indexableObject.uuid = id;
   }
@@ -31,7 +31,7 @@ describe('CommunityAdminSearchResultListElementComponent', () => {
         TranslateModule.forRoot(),
         RouterTestingModule.withRoutes([])
       ],
-      declarations: [CommunityAdminSearchResultListElementComponent],
+      declarations: [CollectionAdminSearchResultListElementComponent],
       providers: [{ provide: TruncatableService, useValue: {} }],
       schemas: [NO_ERRORS_SCHEMA]
     })
@@ -39,7 +39,7 @@ describe('CommunityAdminSearchResultListElementComponent', () => {
   }));
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(CommunityAdminSearchResultListElementComponent);
+    fixture = TestBed.createComponent(CollectionAdminSearchResultListElementComponent);
     component = fixture.componentInstance;
     component.object = searchResult;
     component.linkTypes = CollectionElementLinkType;
@@ -55,6 +55,6 @@ describe('CommunityAdminSearchResultListElementComponent', () => {
   it('should render an edit button with the correct link', () => {
     const a = fixture.debugElement.query(By.css('a'));
     const link = a.nativeElement.href;
-    expect(link).toContain(getCommunityEditPath(id));
+    expect(link).toContain(getCollectionEditPath(id));
   })
 });
