+++ src/app/+admin/admin-search-page/admin-search-results/admin-search-result-list-element/community-search-result/community-admin-search-result-list-element.component.spec.ts
@@ -1,27 +1,27 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { NO_ERRORS_SCHEMA } from '@angular/core';
-import { CollectionAdminSearchResultListElementComponent } from './collection-admin-search-result-list-element.component';
 import { TranslateModule } from '@ngx-translate/core';
-import { TruncatableService } from '../../../truncatable/truncatable.service';
-import { CollectionElementLinkType } from '../../../object-collection/collection-element-link.type';
-import { ViewMode } from '../../../../core/shared/view-mode.model';
-import { CollectionSearchResult } from '../../../object-collection/shared/collection-search-result.model';
-import { Collection } from '../../../../core/shared/collection.model';
+import { TruncatableService } from '../../../../../shared/truncatable/truncatable.service';
+import { CollectionElementLinkType } from '../../../../../shared/object-collection/collection-element-link.type';
+import { ViewMode } from '../../../../../core/shared/view-mode.model';
 import { By } from '@angular/platform-browser';
 import { RouterTestingModule } from '@angular/router/testing';
-import { getCollectionEditPath } from '../../../../+collection-page/collection-page-routing.module';
+import { CommunityAdminSearchResultListElementComponent } from './community-admin-search-result-list-element.component';
+import { CommunitySearchResult } from '../../../../../shared/object-collection/shared/community-search-result.model';
+import { getCommunityEditPath } from '../../../../../+community-page/community-page-routing.module';
+import { Community } from '../../../../../core/shared/community.model';
 
-describe('CollectionAdminSearchResultListElementComponent', () => {
-  let component: CollectionAdminSearchResultListElementComponent;
-  let fixture: ComponentFixture<CollectionAdminSearchResultListElementComponent>;
+describe('CommunityAdminSearchResultListElementComponent', () => {
+  let component: CommunityAdminSearchResultListElementComponent;
+  let fixture: ComponentFixture<CommunityAdminSearchResultListElementComponent>;
   let id;
   let searchResult;
 
   function init() {
     id = '780b2588-bda5-4112-a1cd-0b15000a5339';
-    searchResult = new CollectionSearchResult();
-    searchResult.indexableObject = new Collection();
+    searchResult = new CommunitySearchResult();
+    searchResult.indexableObject = new Community();
     searchResult.indexableObject.uuid = id;
   }
   beforeEach(async(() => {
@@ -31,7 +31,7 @@ describe('CollectionAdminSearchResultListElementComponent', () => {
         TranslateModule.forRoot(),
         RouterTestingModule.withRoutes([])
       ],
-      declarations: [CollectionAdminSearchResultListElementComponent],
+      declarations: [CommunityAdminSearchResultListElementComponent],
       providers: [{ provide: TruncatableService, useValue: {} }],
       schemas: [NO_ERRORS_SCHEMA]
     })
@@ -39,7 +39,7 @@ describe('CollectionAdminSearchResultListElementComponent', () => {
   }));
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(CollectionAdminSearchResultListElementComponent);
+    fixture = TestBed.createComponent(CommunityAdminSearchResultListElementComponent);
     component = fixture.componentInstance;
     component.object = searchResult;
     component.linkTypes = CollectionElementLinkType;
@@ -55,6 +55,6 @@ describe('CollectionAdminSearchResultListElementComponent', () => {
   it('should render an edit button with the correct link', () => {
     const a = fixture.debugElement.query(By.css('a'));
     const link = a.nativeElement.href;
-    expect(link).toContain(getCollectionEditPath(id));
+    expect(link).toContain(getCommunityEditPath(id));
   })
 });
