+++ src/app/+admin/admin-search-page/admin-search-results/item-admin-search-result-actions.component.spec.ts
@@ -2,29 +2,31 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { NO_ERRORS_SCHEMA } from '@angular/core';
 import { TranslateModule } from '@ngx-translate/core';
-import { TruncatableService } from '../../../truncatable/truncatable.service';
-import { CollectionElementLinkType } from '../../../object-collection/collection-element-link.type';
-import { ViewMode } from '../../../../core/shared/view-mode.model';
-import { Collection } from '../../../../core/shared/collection.model';
 import { By } from '@angular/platform-browser';
 import { RouterTestingModule } from '@angular/router/testing';
-import { ItemSearchResult } from '../../../object-collection/shared/item-search-result.model';
-import { ItemAdminSearchResultListElementComponent } from './item-admin-search-result-list-element.component';
-import { getItemEditPath } from '../../../../+item-page/item-page-routing.module';
-import { URLCombiner } from '../../../../core/url-combiner/url-combiner';
-import { ITEM_EDIT_DELETE_PATH, ITEM_EDIT_MOVE_PATH, ITEM_EDIT_REINSTATE_PATH, ITEM_EDIT_WITHDRAW_PATH } from '../../../../+item-page/edit-item-page/edit-item-page.routing.module';
+import { ItemAdminSearchResultActionsComponent } from './item-admin-search-result-actions.component';
+import { Item } from '../../../core/shared/item.model';
+import {
+  ITEM_EDIT_DELETE_PATH,
+  ITEM_EDIT_MOVE_PATH,
+  ITEM_EDIT_PRIVATE_PATH,
+  ITEM_EDIT_PUBLIC_PATH,
+  ITEM_EDIT_REINSTATE_PATH,
+  ITEM_EDIT_WITHDRAW_PATH
+} from '../../../+item-page/edit-item-page/edit-item-page.routing.module';
+import { getItemEditPath } from '../../../+item-page/item-page-routing.module';
+import { URLCombiner } from '../../../core/url-combiner/url-combiner';
 
-describe('ItemAdminSearchResultListElementComponent', () => {
-  let component: ItemAdminSearchResultListElementComponent;
-  let fixture: ComponentFixture<ItemAdminSearchResultListElementComponent>;
+describe('ItemAdminSearchResultActionsComponent', () => {
+  let component: ItemAdminSearchResultActionsComponent;
+  let fixture: ComponentFixture<ItemAdminSearchResultActionsComponent>;
   let id;
-  let searchResult;
+  let item;
 
   function init() {
     id = '780b2588-bda5-4112-a1cd-0b15000a5339';
-    searchResult = new ItemSearchResult();
-    searchResult.indexableObject = new Collection();
-    searchResult.indexableObject.uuid = id;
+    item = new Item();
+    item.uuid = id;
   }
   beforeEach(async(() => {
     init();
@@ -33,20 +35,16 @@ describe('ItemAdminSearchResultListElementComponent', () => {
         TranslateModule.forRoot(),
         RouterTestingModule.withRoutes([])
       ],
-      declarations: [ItemAdminSearchResultListElementComponent],
-      providers: [{ provide: TruncatableService, useValue: {} }],
+      declarations: [ItemAdminSearchResultActionsComponent],
       schemas: [NO_ERRORS_SCHEMA]
     })
       .compileComponents();
   }));
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(ItemAdminSearchResultListElementComponent);
+    fixture = TestBed.createComponent(ItemAdminSearchResultActionsComponent);
     component = fixture.componentInstance;
-    component.object = searchResult;
-    component.linkTypes = CollectionElementLinkType;
-    component.index = 0;
-    component.viewModes = ViewMode;
+    component.item = item;
     fixture.detectChanges();
   });
 
@@ -74,13 +72,8 @@ describe('ItemAdminSearchResultListElementComponent', () => {
 
   describe('when the item is not withdrawn', () => {
     beforeEach(() => {
-        component.dso.isWithdrawn = false;
+        component.item.isWithdrawn = false;
         fixture.detectChanges();
-    });
-
-    it('should not show the withdrawn badge', () => {
-      const badge = fixture.debugElement.query(By.css('div.withdrawn-badge'));
-      expect(badge).toBeNull();
     });
 
     it('should render a withdraw button with the correct link', () => {
@@ -97,24 +90,55 @@ describe('ItemAdminSearchResultListElementComponent', () => {
 
   describe('when the item is withdrawn', () => {
     beforeEach(() => {
-      component.dso.isWithdrawn = true;
+      component.item.isWithdrawn = true;
       fixture.detectChanges();
     });
 
-    it('should show the withdrawn badge', () => {
-      const badge = fixture.debugElement.query(By.css('div.withdrawn-badge'));
-      expect(badge).not.toBeNull();
-    });
-
-    it('should render a withdraw button with the correct link', () => {
+    it('should not render a withdraw button with the correct link', () => {
       const a = fixture.debugElement.query(By.css('a.withdraw-link'));
       expect(a).toBeNull();
     });
 
-    it('should not render a reinstate button with the correct link', () => {
+    it('should render a reinstate button with the correct link', () => {
       const a = fixture.debugElement.query(By.css('a.reinstate-link'));
       const link = a.nativeElement.href;
       expect(link).toContain(new URLCombiner(getItemEditPath(id), ITEM_EDIT_REINSTATE_PATH).toString());
+    });
+  });
+
+  describe('when the item is not private', () => {
+    beforeEach(() => {
+      component.item.isDiscoverable = true;
+      fixture.detectChanges();
+    });
+
+    it('should render a make private button with the correct link', () => {
+      const a = fixture.debugElement.query(By.css('a.private-link'));
+      const link = a.nativeElement.href;
+      expect(link).toContain(new URLCombiner(getItemEditPath(id), ITEM_EDIT_PRIVATE_PATH).toString());
+    });
+
+    it('should not render a make public button with the correct link', () => {
+      const a = fixture.debugElement.query(By.css('a.public-link'));
+      expect(a).toBeNull();
+    });
+  });
+
+  describe('when the item is private', () => {
+    beforeEach(() => {
+      component.item.isDiscoverable = false;
+      fixture.detectChanges();
+    });
+
+    it('should not render a make private button with the correct link', () => {
+      const a = fixture.debugElement.query(By.css('a.private-link'));
+      expect(a).toBeNull();
+    });
+
+    it('should render a make private button with the correct link', () => {
+      const a = fixture.debugElement.query(By.css('a.public-link'));
+      const link = a.nativeElement.href;
+      expect(link).toContain(new URLCombiner(getItemEditPath(id), ITEM_EDIT_PUBLIC_PATH).toString());
     });
   })
 });
