--- src/app/shared/object-grid/admin-search-result-grid-element/item-search-result/item-admin-search-result-grid-element.component.html
@@ -1,41 +1,0 @@
-<ng-template dsListableObject>
-</ng-template>
-<div #badges class="position-absolute ml-1">
-    <div *ngIf="dso && !dso.isDiscoverable" class="private-badge">
-        <span class="badge badge-danger">{{ "admin.search.item.private" | translate }}</span>
-    </div>
-    <div *ngIf="dso && dso.isWithdrawn" class="withdrawn-badge">
-        <span class="badge badge-warning">{{ "admin.search.item.withdrawn" | translate }}</span>
-    </div>
-</div>
-<ul #buttons class="list-group list-group-flush">
-  <li class="list-group-item d-flex justify-content-between">
-    <a class="btn btn-light btn-sm my-1 edit-link" [routerLink]="[getEditPath()]" [title]="'admin.search.item.edit' | translate">
-      <i class="fa fa-edit"></i>
-    </a>
-
-    <a *ngIf="dso && !dso.isWithdrawn" class="btn btn-light btn-sm my-1 withdraw-link" [routerLink]="[getWithdrawPath()]" [title]="'admin.search.item.withdraw' | translate">
-      <i class="fa fa-ban"></i>
-    </a>
-
-    <a *ngIf="dso && dso.isWithdrawn" class="btn btn-light btn-sm my-1 reinstate-link" [routerLink]="[getReinstatePath()]" [title]="'admin.search.item.reinstate' | translate">
-      <i class="fa fa-undo"></i>
-    </a>
-
-    <a *ngIf="dso && dso.isDiscoverable" class="btn btn-light btn-sm my-1 private-link" [routerLink]="[getPrivatePath()]" [title]="'admin.search.item.make-private' | translate">
-      <i class="fa fa-eye-slash"></i>
-    </a>
-
-    <a *ngIf="dso && !dso.isDiscoverable" class="btn btn-light btn-sm my-1 public-link" [routerLink]="[getPublicPath()]" [title]="'admin.search.item.make-public' | translate">
-      <i class="fa fa-eye"></i>
-    </a>
-
-    <a class="btn btn-light btn-sm my-1 delete-link" [routerLink]="[getDeletePath()]" [title]="'admin.search.item.delete' | translate">
-      <i class="fa fa-trash"></i>
-    </a>
-
-    <a class="btn btn-light btn-sm my-1 move-link" [routerLink]="[getMovePath()]" [title]="'admin.search.item.move' | translate">
-      <i class="fa fa-arrow-circle-right"></i>
-    </a>
-  </li>
-</ul>
